+++ poem-openapi/src/payload/mod.rs
@@ -9,11 +9,7 @@ mod response;
 use poem::{Request, RequestBody, Result};
 
 pub use self::{
-    attachment::Attachment,
-    binary::{Binary, BinaryStream},
-    json::Json,
-    plain_text::PlainText,
-    response::Response,
+    attachment::Attachment, binary::Binary, json::Json, plain_text::PlainText, response::Response,
 };
 use crate::{
     registry::{MetaSchemaRef, Registry},