+++ poem-openapi/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - Add `remote` attribute to `Enum` macro.
+- Remove the `BinaryStream` type, use `poem::Body` instead.
 
 # [1.1.1] 2021-12-12