+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.2.7] 2021-12-19
+
+- Make the `OpenAPI` macro can now report duplicate routing errors.
+
 # [1.2.4] 2021-12-18
 
 - Fix the parameter validator will cause compilation failure in some cases.